django/pyproject.toml
[project]
name = "timerecorder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asgiref==3.8.1",
    "django==5.2.1",
    "django-cors-headers==4.7.0",
    "django-extensions==4.1",
    "django-filter==25.1",
    "django-import-export==4.3.7",
    "djangorestframework==3.16.0",
    "djangorestframework-jsonapi==7.1.0",
    "djangorestframework-simplejwt==5.5.0",
    "inflection==0.5.1",
    "psycopg==3.2.9",
    "python-dotenv==1.1.0",
    "pytz==2025.2",
    "requests==2.32.3",
    "sqlparse==0.5.3",
    "tzdata==2025.2",
]
